[SC-4521] Move the shared daemon test doubles out of the ideation tests
fakeRunner, fakeCreator and fakeEditor are used by the descedit tests too, and
ideation_test.go is about to be deleted. No production change.
